+++ .gitignore
@@ -1,0 +1,4 @@
+# ignore not needed files
+
+.idea/
+*.pyc
