+++ build_square.py
@@ -1,0 +1,16 @@
+# Вывести прямоугольник со сторонами a,b. Например a = 3, b=2
+
+
+def print_rectangle(n, m):
+    for i in range(1, n + 1):
+        for j in range(1, m + 1):
+            if (i == 1 or i == n or
+                    j == 1 or j == m):
+                print('*', end='')
+            else:
+                print(' ', end='')
+
+        print()
+
+
+print_rectangle(3, 4)
