+++ copy_list_to_function.py
@@ -1,0 +1,9 @@
+# Как передать копию списка или словаря в функцию?
+
+
+def get_list(list):
+    return list
+
+
+array_1 = [1, 2]
+print(get_list(array_1.copy()))
