+++ find_matrix_sum.py
@@ -1,0 +1,8 @@
+# Найти сумму элементов матрицы
+
+summation = 0
+matrix = [[1, 2, 3], [4, 5, 6], [7, 8, 9], [8, 7, 6]]
+for raw in matrix:
+    for value in raw:
+        summation += value
+print(sum)
